release/1.0.2 - corrected access to database

+++ python/reset_database.py
@@ -1,16 +1,29 @@
 #!/usr/bin/env python3
+
 import subprocess
 import os
 import sys
 from pathlib import Path
 from typing import Optional
+#local
+# NAME="symfony"
+# USER="root"
+# PASS="root"
+# HOST="127.0.0.1"
+# PORT="3308"
+#live
+NAME="kwadro_laravel"
+USER="kwadro_laravel"
+PASS="y22KN_t+u8"
+HOST="kwadro.mysql.tools"
+PORT="3308"
 
 def reset_database(
-    db_name: str = "symfony",
-    db_user: str = "root",
-    db_pass: str = "root",
-    db_host: str = "127.0.0.1",
-    db_port: str = "3308",
+    db_name: str = NAME,
+    db_user: str = USER,
+    db_pass: str = PASS,
+    db_host: str = HOST,
+    db_port: str = PORT,
     migrations_dir: Path = Path(__file__).parent.parent/ "migrations",
     confirm: Optional[bool] = True
 ) -> None:
@@ -32,6 +45,7 @@ def reset_database(
         return
     PROJECT_ROOT = Path(__file__).parent.parent.resolve()
     SCRIPT_PATH_IMPORT = PROJECT_ROOT / "bash" / "import-entity.sh"
+    DROP = 0
     # --------------------------
     # CLEAN CACHE
     # --------------------------
@@ -45,30 +59,32 @@ def reset_database(
     # --------------------------
     # DROP & CREATE DATABASE
     # --------------------------
-    print("🗑 Dropping ALL tables...")
-    mysql_command = (
-        f"SET FOREIGN_KEY_CHECKS = 0; "
-        f"DROP DATABASE IF EXISTS `{db_name}`; "
-        f"CREATE DATABASE `{db_name}`;"
-    )
-    try:
-        subprocess.run(
-            [
-                "mysql",
-                f"-u{db_user}",
-                f"-p{db_pass}",
-                f"-h{db_host}",
-                f"-P{db_port}",
-                "-e",
-                mysql_command
-            ],
-            check=True
+    if(DROP==1):
+        print("🗑 Dropping ALL tables...")
+        mysql_command = (
+            f"SET FOREIGN_KEY_CHECKS = 0; "
+            f"DROP DATABASE IF EXISTS `{db_name}`; "
+            f"CREATE DATABASE `{db_name}`;"
         )
-    except subprocess.CalledProcessError:
-        print("❌ Failed to drop/create database")
-        sys.exit(1)
+        try:
+            subprocess.run(
+                [
+                    "mysql",
+                    f"-u{db_user}",
+                    f"-p{db_pass}",
+                    f"-h{db_host}",
+                    f"-P{db_port}",
+                    "-e",
+                    mysql_command
+                ],
+                check=True
+            )
+        except subprocess.CalledProcessError:
+            print("❌ Failed to drop/create database")
+            sys.exit(1)
 
-    print("✔ Database recreated")
+        print("✔ Database recreated")
+
 
     # --------------------------
     # CLEAN MIGRATIONS
